Ndr/add rtree

+++ rust/compass-core/src/model/graph/vertex_id.rs
@@ -3,7 +3,7 @@ use std::{cmp::Ordering, fmt::Display};
 use serde::{Deserialize, Serialize};
 
 #[derive(Copy, Clone, Serialize, Deserialize, PartialEq, PartialOrd, Eq, Hash, Debug, Default)]
-pub struct VertexId(pub u64);
+pub struct VertexId(pub usize);
 
 impl Ord for VertexId {
     fn cmp(&self, other: &Self) -> Ordering {
